pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"fastapi",
	"uvicorn[standard]",
	"httpx",
	"python-dotenv",
	"google-genai",
	"sqlalchemy>=2.0",
	"psycopg2-binary",
	"bcrypt",
	"PyJWT",
	"neo4j>=5.0",
	"celery",
	"pypdf",
	"qdrant-client",
	"slowapi"
]
